Add files via upload
Full Main method frame complete. Switch/case structures added for pre-Flop and post-Flop stages, with the latter having a nested For loop accounting for the identical possible actions in the Flop, Turn, and River stages of the game.

Showdown method is called after the completion of the For loop, and then the game's While loop enters a new iteration.

Whitespace has been standardized for readability of the code and the printed statements. When more code becomes available, I will take a look to standardize formatting across all pushes.

Header file not included because no changes have been made.